refactor: reorganize Events/Heartbeat/AgentStatus into Scheduler+Logs
Reflects the event-driven architecture: Heartbeat and Cron are both
CronEngine consumers on the same dispatch bus (→ Scheduler page).
Event Log and Tool Call Log are parallel audit trails (→ Logs page).

Old URLs (/events, /heartbeat, /agent-status) redirect to new pages.
