+++ ui/src/components/Sidebar.tsx
@@ -46,36 +46,27 @@ const NAV_SECTIONS: NavSection[] = [
           </svg>
         ),
       },
-      {
-        page: 'events',
-        label: 'Events',
-        icon: (active) => (
-          <svg width="18" height="18" viewBox="0 0 24 24" fill={active ? 'currentColor' : 'none'} stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round">
-            <polyline points="22 12 18 12 15 21 9 3 6 12 2 12" />
-          </svg>
-        ),
-      },
     ],
   },
   {
     sectionLabel: 'Agent',
     items: [
       {
-        page: 'agent-status',
-        label: 'Agent Status',
+        page: 'scheduler',
+        label: 'Scheduler',
         icon: (active) => (
           <svg width="18" height="18" viewBox="0 0 24 24" fill={active ? 'currentColor' : 'none'} stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round">
-            <rect x="4" y="4" width="16" height="16" rx="2" />
-            <path d="M4 12h4l2-4 4 8 2-4h4" />
+            <circle cx="12" cy="12" r="10" />
+            <polyline points="12 6 12 12 16 14" />
           </svg>
         ),
       },
       {
-        page: 'heartbeat',
-        label: 'Heartbeat',
+        page: 'logs',
+        label: 'Logs',
         icon: (active) => (
           <svg width="18" height="18" viewBox="0 0 24 24" fill={active ? 'currentColor' : 'none'} stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round">
-            <path d="M20.84 4.61a5.5 5.5 0 0 0-7.78 0L12 5.67l-1.06-1.06a5.5 5.5 0 0 0-7.78 7.78L12 21.23l8.84-8.84a5.5 5.5 0 0 0 0-7.78z" />
+            <polyline points="22 12 18 12 15 21 9 3 6 12 2 12" />
           </svg>
         ),
       },
